requirements.txt
requests>=2.31.0
pandas>=2.1.0
pyarrow>=14.0.0
pyyaml>=6.0
hyperliquid-python-sdk>=0.23.0
scipy>=1.11.0
websockets>=12.0
zstandard>=0.22.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
